initial commit of offset-based getter
The idea here is to try to remove "cast_to_layout_compatible", which
technically violates strict aliasing rules. Instead, we want to try
to compute something like "offsetof" at compile-time for each struct member,
and use pointer arithmetic and reinterpret casts to do this legally.

The basic idea is to take the function "make_flat_tuple_of_references"
and generalize it, so that it need not actually work with tuples.
Instead, it can work with any class type, provided that it is also given
a "getter" function object which it can use to do "get<idx>".

We provide a trivial getter, which calls sequence_tuple::get.

We also provide a "offset_based_getter" which can work with user-defined
structure types, provided we are given a layout-compatible tuple as a template
parameter.

The basic idea is to take the layout-compatible tuple, then replace all the
members with corresponding `std::aligned_storage_t`. This doesn't change the
layout, but it makes everything trivially constructible and constexpr and all that.

Then we can construct that on the stack, and take differences between the main tuple
address, and the members' addresses.

At time of writing the tests are passing for me, but I haven't tested with compilers
other than gcc, or examined assembly to see if this method is being optimized the
way I hope it is.

This commit changes both core14_classic and core14_loophole to use the offset_based_getter,
but until further testing it might have been better to keep the other method around also.